backend/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-cors-middleware==1.3.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
Faker==30.3.0
idna==3.10
inflection==0.5.1
packaging==24.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
setuptools==75.1.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
mysqlclient==2.2.4
python-dotenv==1.0.1
drf-nested-routers==0.91.0
rdflib==7.0.0

